+++ src/tile/utils.rs
@@ -1,25 +1,29 @@
-use bevy::{prelude::*, render::camera::RenderTarget};
+use bevy::{
+    prelude::*,
+    render::camera::RenderTarget,
+    window::{PrimaryWindow, WindowRef},
+};
 
-pub fn get_window_width_height(windows: &Res<Windows>, camera: &Camera) -> Option<Vec2> {
-    let wnd = if let RenderTarget::Window(id) = camera.target {
-        windows.get(id)?
+pub fn get_window_width_height(
+    windows: &Query<(Entity, &Window, Option<&PrimaryWindow>)>,
+    camera: &Camera,
+) -> Option<Vec2> {
+    let wnd = if let RenderTarget::Window(WindowRef::Entity(entity)) = camera.target {
+        windows.iter().find(|(e, _, _)| *e == entity)?.1
     } else {
-        windows.get_primary()?
+        windows.iter().find(|(_, _, p)| p.is_some())?.1
     };
 
     Some(Vec2::new(wnd.width(), wnd.height()))
 }
 
-pub fn get_map_width_height(camera: &Camera, transform: &GlobalTransform) -> Vec2 {
+pub fn get_map_width_height(camera: &Camera, transform: &Transform) -> Vec2 {
     let (left, top, right, bottom) = get_map_coords_of_edges(camera, transform);
     Vec2::new(right - left, bottom - top)
 }
 
 // https://bevy-cheatbook.github.io/cookbook/cursor2world.html
-pub fn get_map_coords_of_edges(
-    camera: &Camera,
-    transform: &GlobalTransform,
-) -> (f32, f32, f32, f32) {
+pub fn get_map_coords_of_edges(camera: &Camera, transform: &Transform) -> (f32, f32, f32, f32) {
     let ndc_tl = Vec2::new(0.0, 0.0) - Vec2::ONE;
     let ndc_br = Vec2::new(2.0, 2.0) - Vec2::ONE;